+++ plugins/paragraphs/builds/compromise-paragraphs.js
@@ -152,9 +152,9 @@
       // returns boolean
       has: function has(str) {
         return this.paragraphs.some(function (docs) {
-          for (var i = 0; i < docs.length; i++) {
-            return docs[i].has(str);
-          }
+          return docs.some(function (doc) {
+            return doc.has(str);
+          });
         });
       },
       //loops